+++ lib/model/named_api_resource_list.g.dart
@@ -1,21 +1,25 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'pokemons.dart';
+part of 'named_api_resource_list.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-Pokemons _$PokemonsFromJson(Map<String, dynamic> json) => Pokemons(
+NamedAPIResourceList _$NamedAPIResourceListFromJson(
+        Map<String, dynamic> json) =>
+    NamedAPIResourceList(
       json['count'] as int,
       json['next'] as String?,
       json['previous'] as String?,
       (json['results'] as List<dynamic>)
-          .map((e) => Pokemon.fromJson(e as Map<String, dynamic>))
+          .map((e) => NamedAPIResource.fromJson(e as Map<String, dynamic>))
           .toList(),
     );
 
-Map<String, dynamic> _$PokemonsToJson(Pokemons instance) => <String, dynamic>{
+Map<String, dynamic> _$NamedAPIResourceListToJson(
+        NamedAPIResourceList instance) =>
+    <String, dynamic>{
       'count': instance.count,
       'next': instance.next,
       'previous': instance.previous,