+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [ ] make api crawler poke api
 - [ ] make screen search pokemon
 - [x] make screen list berry
+- [x] make screen berry detail
 
 ### 27-12-2021
 - [x] init project